License the bulk of the project under MIT
We've been working without a license for a while now. We have assets
that have explicit licenses, but the unlicensed nature of the code and
self-produced assets was lending itself to possible confusion and
conflict down the road.

We should explicitly state the licensing situation in the forematter
of the README. In that case, we will need an actual license to
contrast with.

Looking over the various licenses available, I chose MIT. I felt the
simplicity and freedom offered by MIT would best fit the spirit of the
project.