Update the primary project README
The primary project README was a single line that just said something
to the effect of "A Godot Project".

As part of our overall documentation push, I figured this would be a
good starting point. So the primary README now explains why the
project exists, the particular peculiarities of our documentation, a
short history about WHY Godot of all things, short instructions on
running the project, and a note on the unusual licensing of this
project.

Most importantly, this update establishes a good standard for our
documentation: it is not enough to explain only what. We must also
explain WHY.