Update the color documentation
After updating the project root documentation, I realized that the
colors documentation was missing a couple of critical items. It
explained what the files were, and why we had them.

However, it didn't explain why we felt the need to have a palette to
begin with, or WHY we used that particular palette. So I added a
couple of paragraphs explaining why we were using Resurrect64.

I suppose it doesn't matter that much, but I think it's important to
help enforce continuity with existing art assets.