Redo the physics layers
I was preening the physics layers during our documentation push when I
suddenly realized an error that had thrown the whole configuration out
of whack.

Both navigation meshes had struggled with creating meshes under
terrain objects. I assumed this was because the nav mesh baking
algorithms were bad. Turns out that's only partly correct. See, I was
setting the agent height to match the effective height of our pawn:
1.61. FOR SOME REASON, that was preventing the nav mesh generation
from passing under doorways. If I halve it, the problem gets fixed.
Maybe it's an extent-height terminology issue?! I DON'T KNOW!

Point is, the whole three-kinds-of-terrain paradigm had been based on
the flawed assumption that something like that was necessary. Turns
out we just needed to modify the navigation baking algorithm.

That meant I had to go back over every scene that involved a physics
scene (A LOT) and make sure they were right.

I also preened a lot the speculative layers off - we can capture layer
speculation in the documentation. That's part of the reason it should
be there: to remove such speculative info from the game proper.

As a sign of confidence, I also moved the "Camera Obstruction" layer
to the back of the layers. That way we won't have to keep moving it
up.