Update the column's primary material exports
Previously, we were using a series of enums to control whether the
individual faces of the columns displayed the primary material or
the alternate material.

That was okay, but I felt we could do better. A more compact solution,
codewise, would be to use bit-packing. So I tried it, and I'm pretty
happy with the result.

This necessitated changes in the Wall set.