Document the Motion AI directory
Added documentation to the AI directory. One of the more important
items was that I developed a sort of documentation standard: for
important scenes, we give a breif description, followed by an
enumeration of the configurables, (public) variables, and signals.
Public variables is used to differentiate from private variables
that really shouldn't be messed with.

This is different from how we documented stuff in the buildings
directory - I think I'll use this more intense documentation in
cases where the effects of configurables aren't immediately obvious.
Or, better yet, where programmatic integration is a necessity.

During documentation, I noticed that our resources were named poorly.
If you try to document something and you notice it doesn't make all
that much sense, best to change it. So I changed the names of most of
our pawn assets.

In recognition of only using the alternate colors in tests, I
relocated the alternate sprite color schemes to the test directory.

I noticed that the order of things in BasePawn.gd was out of whack, so
I rearranged things. I also removed the old pawn_died signal. We'll
probably reimplement that eventually, but for now it was a relic.