Edit Pawn sprites directly instead of overriding
I had added an "Override Sprite Frames" configurable to the Pawn so
that we could configure a different sprite in the editor to appear at
runtime.

However, I realized that, in code, we would basically have two ways of
doing this: either using our easy configurable or reaching down into
the Pawn and changing the sprite manually. In addition to this
ambiguity, I realized that the whole configurable was moot to begin
with since Godot has the "editable children" option on instanced
scenes.

You could make the case that this feature should have stayed for
convenience, but I don't want the project to get bloated. With that in
mind, I removed the feature from Pawn and the documentation. The
Viewport Shader test was updated so that we use the various Pawn
sprites.