Use Godot native Navigation for shader testing
It was bugging me that we used a dynamic navigation mesh for testing
out shaders. It was mixing up our different tests - technically, our
Viewport Shader Test was testing both shaders and the dynamic meshes.

That mixing of types was bugging me, so I switched to just using a
regular Navigation mesh.