Document the "scenes" directory
I removed unnecessary/outdated scenes from the scenes directory, and
then documented what was left. That was all of one scene - that
CameraRig. While a complex scene, it wasn't all that difficult to
document.

I did rearrange a couple of things in the CameraRig and, subsequent to
that, the Viewport Shader Test.