Update silhouette shaders ahead of documentation
I was starting to write documentation for the shaders when I realized:
the silhouette shaders were poorly structured.

Firstly, the texture input for the silhouette shaders wasn't a texture
uniform, so the name was something like "Tex Frg 04". That wasn't very
documentation friendly, so it needed to be replaced with a "Viewport
Texture" Uniform value.

Secondly, the color black was being eliminated from the sprites as
alpha. It occurred to me that black will probably be a pretty common
color for sprites, so I should REALLY do something about that.

That meant changing the Silhouette camera with a brand new Environment
resource - the SilhouetteWorldEnvironment. The key difference between
this environment and our SceneDefaultEnvironment is the
hot-pink/magenta background color (#ff00ff).

With the background color unusually bright and distinct, I added a new
color uniform to both Silhouette shaders - the "Alpha Color". This
controls what color gets alpha'd out, just in case we have a desparate
need to use magenta in a sprite.

This makes our silhouette shaders much easier to work with and
understand.