Document the terrain directory
Documented the terrain directory and it's subdirectories. Except I
didn't document just the "terrain" directory. I tried to document that
directory, but there was nothing to say that wasn't repeated and
elaborated on in subsequent directories.

I deleted the Rock scene as well - like many other scenes removed in
the documentation push, the only thing I could think to say about it
was "I should remove this". The multiple collision models was actively
interferring with pathing. Rocks will return - one day.

Sizing up what's left - it looks like we got about 5 directories left
to document. It'll be a tall order but I think we can slice it.