Revise the Viewport Shader Test
Previously, the Viewport Shader Test involved moving the pawns around
and hoping that they crossed each other at some opportune moment. It
was literally random movement hoping for a random result.

Problem is, the shaders were behaving strangely on certain surfaces
and around each other. I needed a consistent test and all I had was a
bunch of ants running around.

So, I've revised the test. The pawns now cycle, one at a time, through
a series of deliberately designed tests. I was able to save a lot of
code and a lot of headache by using yield() functions. It's still a
lot of code, but not as bad as it could've been.