Use HidingMesh for Fill-Multiply Viewport Shaders
Previously, the meshes we used for testing the fill and multiply
viewport shaders we're always enabled. This was mostly just laziness
on my part.

It was starting to affect the shader tests, because the shader always
being on was leading to some situations where I didn't know EXACTLY
what I was supposed to be seeing.

These tests need some sort of clarity, even if minimal. To correct
this, we now us the HidingMesh scene for the Multiply-Fill shaders
(instead of the generic meshes we were using before).