Band-aid the KinematicDriver position loop error
I've been experiencing an occassional issue with the Pawns where,
every once in a while, the Pawn would get stuck and rapidly oscillate
in whatever position it was in and refuse to path further.

Some investigation found that the Pawn was consistently moving past
it's target, reversing course, and then moving past the target again.
WHY that was happening is beyond me. I've taken precautions to ensure
that KinematicDriver bodies only move an exact distance, yet here we
were rubber banding around the goal. I currently suspect some sort of
interplay between the navigation mesh (giving us a bad value), the
collision model (the Pawn's box catching on something), and the
KinematicDriver's floating and gravity (causing us to constantly rise
and fall on an extremely small scale).

At this moment, the interplay between all three is too much for me to
try and somehow patch and smooth over, so I instead decided to just
cope with whenever the KinematicDriver has a problem. To that end,
there's a new signal - error_microposition_loop. I termed the error as
a "microposition loop" on account of the loop's small position
variance. It will be up to scenes that integrate KinematicDriver to
handle this error. For the Pawn, that just meant clearing the
KinematicDriver's current goal.

This meant we could knock the Pawn's KinematicDriver Goal Tolerance
down to a lower level - more accurate movement! To help avoid this
issue in the future, the Pawn's actual collision model has been
changed to a cylinder. The curves of the cylinder help to avoid
catching on edges, which MIGHT have been contributing to the issue.

The error checking was implemented using a history track-and-check
method. Because the KinematicDriver records new values so quickly, the
error doesn't trigger when the Pawn is stuck in a different way - like
trying to walk up a slope and sliding down.

However, I am worried that this method might fire off at some
incorrect moment. We'll keep an eye on it...