Switch to using a capsule for the Pawn
After the last update, we had changed the Pawn's collision model to a
cylinder - the theory being the rounded edges would stop the body from
catching and creating microposition errors.

However, we were still getting some catches. I was trying to avoid
using capsules because the curved bottom is hard to work with (it can
be pretty fiddly). However, since the cylinder wasn't working, using a
capsule was the only solution I could see.

So the Pawn now uses a capsule collision shape. Because the bottom of
the capsule is curved, the Pawn was able to step up unto ledges over
0.5 units. To fix this, we needed to mess with the size of the capsule
as well as the configured float height. The step up estimate is now
much more rough, but it can at least clear steps .5 units high.

I also added new to-do items, but I've added them to the back. I've
procastinated long enough. Next step: weapons!