Add an OpenSCAD sword script
In order to start working with attacks, we need a whole suite of
things. We'll need animations and weapon sprites. And to make those,
we'll need a weapon model.

To model the weapon, we've added a new OpenSCAD script. This script
creates a sword of... indeterminate type. It should really be sword of
the late middle ages, or the early Renaissaince. Instead, it's a mutt
of swords from the Migration Period through the Viking Age. But my
research has really turned up that, more than anything, swords are
imprecise and fluid in their form. So it doesn't really matter. What
matters most of all is that it LOOKS like a sword.

Also, removed the Rothenberg template house. We've had the procedural
building components for a while. There is no need for that old script
now.