Add in a sword swinging animation (and some more)
After creating a sword, the natural next step was to create a sword
swinging animation.

Unfortunately, that wasn't as wasy as it seemed. Firstly, we needed a
consistent way to position the sword model - this lead to the creation
of the "prepare_accessories" script. I hope to expand that script as
we build more accessories.

However, the sword didn't quite fit the Pawn's hand, which meant I had
to go and change sword generation script to increase the length of the
grip.

I then realized that we had a naming problem: all of these scripts
were Pawn specific, but were generally named. So I renamed a lot of
scripts to include some sort of "pawn" identifier.

I also realized that we were going to end up with a LOT of poses and a
LOT of animations. So I made some changes - firstly, what was going to
be our library of poses is now the "core" library of poses, and mostly
just has default stuff pose functions. I also created a new directory
specifically for animation scripts. Pose functions peculiar to an
animation should go in the associated script in the animation
directory.

I added two animations - the basic walk cycle and righthand horizontal
sweep (meant to be used with the sword or any weapon).

Completing the animation led to a ANOTHER NEW problem where the sword
was peeking out of the frame. This required us to increase the size of
the sprites (by adding extra transparent space). And THAT required us
to update the sprite frames appropriately.

Next, we need to find a way to integrate the sprite frames with the
weapons. I THINK I have a way of doing that...