Add a compositing configuration node
I developed an arragement of compositing nodes that allows us to
easily render only the weapon portion of a sprite, effectively masking
out the Pawn. However, I didn't yet have a way of doing it
consistently, from the ground-up.

The technique itself involves assigning every object a dedicated "pass
index", and then masking in only the items of a certain pass index.
This seems to be a feature that's entirely unique to the 'Cycles'
render engine in Blender.

I added a script to enable/disable the weapons masking. It features a
GUI, prompting the user for the current configuration action. Neat!
The introduction of the "pass index" concept necessitated reciprocal
changes in several of the other scripts.

However, I'm not happy with this script. It presents three modes -
render both, weapon-only mask, and pawn-only mask. In reality, there
should be three modes: none, only pawn (as in, shows the whole pawn,
no weapons or masking), and weapon-only mask. Right now I'm only
getting the Pawn-only by moving out the WHOLE weapon sprite. There's
got to be a better way.