Add a modal dialog choice to the accessory prep
Previously, the Accessory Preparation script required the user to edit
the script by hand and do some manual preparation ahead of time so
that the script would actually work.

That really went against our overall policy of making self-contained
scripts. So, I added a modal dialog to the script that allows us to
specify WHAT object should be prepared as an accessory, and WHICH kind
of accessory it should be prepared as.

It works pretty well. You still have to load the model ahead of time
BUT other than that it is very self contained.