Modify the Pose script to include a Pose GUI
For a while, we've had multiple scripts where the user was required to
edit the files and add or remove function calls to get the results
they wanted. Part of the reason we have these scripts is for
consistency and encapsulation - the manual editing method encouraged
neither.

So, I added a GUI to the Pawn Pose (Core) script. If run as the main
script, it now summons a modal dialog that allows you to manually pose
the Pawn. This allows the script to double up as a recovery script of
sorts (in that it resets us to default).