Experiment with integrating weapons into the Pawns
While we were developing a solid base for creating weapon sprites
consistently, I was also slowly experimenting with integrating the
weapon sprites into the Pawns.

The first change I made was to get rid of the "pawn revision" and
"color" naming distinctions - these were complicating the naming
patterns. Different color pallettes should have semantic names, not
color numbers (especially since we're gonna go with a more homogenous
Team-A-vs-Team-B prototype). In addition, there's no real reason to
track the revision of the pawn when we only have one version and we
have Git version control in place.

Spritesheets are now identified as [subject]_[animation].png - with
one spritesheet per animation. Weapon spritesheets are to be placed in
a "weapons" subdirectory with the relevant Motion AI, since they are
specific to the Motion AI type. The weapons take the form of
[subject]_[weapon type]_[animation].png.

I also decided that a simplification of that animation names was in
order. Going [direction]_[animation] lead to the frames being spread
out; it was better to do the frames as [animation]_[direction].

Also, I found out that the KinematicDriver has apparently just been
mirroring west-facing animations to make east-facing animations THIS
ENTIRE TIME! I must have done that back when I was using Surt's
sprites which was like... April? March? Sometime at that time in 2020.
I guess it's a testament to the technique that I didn't notice.