Update default position poses
I had avoided using several default pose functions - the ones for the
hands, feet, and body - because they were positioning things wrong. I
was under the impression that this was a glitch (i.e. not my fault).

As it turns out, I was totally wrong. It WAS all my fault! It seems
like I really didn't take the time to think out the default location
of a lot of the body parts. The most common issue was a failure to
account for moving the origin is the Pawn Prepare script. Changing the
default location values greatly improved the default functions to make
them ACTUALLY usable.

Because the default functions all work now, I was able to add an all
encompassing default function, and then call that in the animation
scripts.

Also, I noticed that the README file excluded the Clear Animation
script because... I must have forgotten, I guess. So that got added
in.