Add in the sweep, chop, and stab animations
Put in the three new weapon animations into the actual frames
resources. The question now - where do we go from here?

We need to suss out the relationship between a motion AI, the weapon
it equips, and it's available attacks. Who manages the sprites? Who
decides when to attack? What triggers the animation? How is damage
logged? This is gonna take a fair bit of work.