Begin basic weapon integration
Adding weapons is gonna be a huge step forward for us - unfortunately
it's also gonna be pretty complex and probably require a lot of
revisions to existing systems.

That's why I am, to the fullest extent possible, trying to take very
small super-baby babysteps. So, to start, I've added a
WeaponAttackTest and an equipment configurable. That's all. Oh, and I
renamed the short sword animation frames (I figured it was worth it to
specify what the resource WAS).

It occurs to me now that it will be difficult to integrate an attack
with the KinematicDriver. We might have to tear it apart and build
something new...