Update the to do list to reflect sword integration
I've been putting off updating the to do list, but it was starting to
look like we needed a reappraisal of our objectives.

That's mostly because we need to thoroughly develop the sword before
we move onto the other weapons. We've developed a pathway for creating
the sprites, but now we need to create a top-to-bottom way of handling
attacks. That includes the AI around the attacks - it's gonna be a
whole lotta work.