Update the Icon
I know it wasn't anything we were trying to work on, but the icon was
bugging me. I have a lot of Godot projects in my project manager
(mostly samples and feature references) and just using the bottom-bin
basic Godot icon was actually starting to annoy me.

It only took me about thirty minutes or so to whip up an icon using
two squares and the Pawn sprite. Easy, and distinctive.