Add the Extended State Machine plugin
As proposed in the to do list, we're gonna try experimenting with the
state machines.

In order to accomplish that, I downloaded Etienne Blanc's Extended
State Machine (XSM) plugin. The XSM provides a lot of extra utility
for state machines and will probably be our best bet, plugin wise.

As I outlined in the README, there is ONE issue I have - the state
machines only operate in the _physics_process function. We should see
if there's a way to have each state work with _process and
_physics_process in parallel.