Create a KinematicDriverMachine to parallel Driver
In order to test out how the Extended State Machines addon works, I
figured it would be best to start by reimplementing something
familiar: the KinematicDriver.

The biggest issue I ran into is that the Extended State Machine is
very clearly designed as an encapsulated thing. Passing information
between the states to a lot of sussing out. Eventually, I settled on a
model of Cores and Machines. In this model, Cores serve as information
stores and configuration vehicles. The machines handle the complex
execution logic. It's a bit of a messy solution, but the integration
is easy and the configuration items are nice and explicitly separated.

Note that all of this is still very much subject to change. As time
goes on, I suspect that the Cores will lose their data role in favor
of that information going to the machine's root state. Eventuall, the
cores might just be information vehicles. Maybe...

So, we now have two new scenes: a KinematicCore (which provides and
tracks Kinematic data and config items) and the corresponding
KinematicDriverMachine.

I didn't integrate it into the Pawn - yet! We'll handle that over the
next few commits. Instead, I just integrated it into the CubeDriver
test, which was meant as a demonstartion of integrating the
KinematicDriver.