Move status variables from Core to Machine
Originally we had the status of Kinematic items - i.e. the current
orientation, the current target - attached to the KinematicCore. The
idea was that this would provide a common status interface across all
of the Machines.

However, the more I thought about it, the more I disliked this idea.
As state machines, the Machines must encapsulate state to the highest
degree possible. To have the state spread across multiple items was
just kinda weird. I also realized that what "states" are shown to the
outside world should be the choice of the machine. Anything
integrating the machine has to deal with the machine's interface; not
the reverse.

So, I moved several of the KinematicCore's variables to the
KinematicDriverMachine. I'm hopinh this proves sustainable.