Add GraphViz scripts for documentation
The Extended State Machines are pretty complicated. They're gonna need
extensive documentation so we know what's going on.

Of course, the best way to do that would be through diagrams. However,
trying to design edge-and-node graphs by hand can prove absolutely
nightmarish. Plus, that would violate our principle of
reproducability. The solution, then, would be scripts that allow us to
design a graph via text/source and then have a program output that
image.

And GraphViz is the solution! GraphViz is free and easy-to-use, with a
well-extant body of documentation.

To test out it's capability, I added scripts for documenting the KDM
and then updated the associated documenation. I'm pretty happy with
it; next comes documenting the Extended State Machine addon for any
newcomers.