Document how XSM works so people know what's up
When I was starting with XSM, it took about two-to-three days of me
staring at the sample code to understand what was happening. While it
was actually pretty easy in retrospect, I don't want anybody to get
hung up on it - ESPECIALLY since I'm going to be using the XSM a lot
more.

So, I took a day and wrote some documentation for the Extended State
Machines, including the background of StateCharts and regular state
machines. Maybe a little overzealous but I think extra documentation
is always an improvement over little documentation.