Convert TestShell to use XSM
Two scenes made use of states in a way that, point-perfect, gave me a
headache: TestShell and SlopeStepTest. So, before trying to embark on
ripping apart the Pawn class, I decided to cut my teeth further by
using these two scenes.

The TestShell was an easy convert, and a massive improvement. The
SlopeStepTest... not so much. I guess I'm still trying to figure out
what's the best way to do stuff. The TestShell was a good
demonstration of how to integrate signal listening into different
states.

The only thing I really dislike is how bothersome it is to access the
nodes we need from states deep in the tree. You can't use the very
beautiful and clear $ notation; you have to resort to get_node().
I guess I'll have to think of a better way...