Completely replace KinematicDriver with the Machine
I have removed the KinematicDriver, replacing it with the
KinematicDriverMachine. Hopefully, this will prove to be a more
malleable solution.

I had to make several changes - the "BasePawn" class has been changed
into just the sprites, collision, and KinematicCore. This will serve
as the basis for other iterations of the Pawn in the future. The
common collision and Kinematic elements allow us to test features
piecemeal.

To demonstrate this concept, we have two derived forms of Pawn. The
first is the SpriteDiagnosticPawn - a constant issue we had was
testing out how certain animations looked with a given kinematic
profile. With the old pawn, there was no way to set animations
manually, so the Pawn Standing test relied on a hacky work around.

This required an additional machine that does NOTHING but fall down.
That is the brand new Falling Body Machine. You can see it as almost
a regression of the logic in the KinematicDriverMachine.

The next Pawn is similar to the BasePawn that we started with. It uses
the KinematicDriverMachine to move the Pawn around, as perdicted. I
did run into a couple of issues while integrating. For compatability,
the KinematicDriverMachine (and KinematicPawn) now accept both
individual targets and target paths (i.e. Arrays). The KinematicDriver
will work through any path points, and the current target, before
sending out a "complete path" signal.

BUT - does this all work well? Uh... no, not exactly. There are two
key issues.

First, I've seen major frame loss, especially with the KinematicPawns.
That's because we update on a signal emission from the Kinematic
Driver Machine, which is generally emitted EVERY SINGLE TIME. That's
not going to work. We're gonna need to develop an intermediate
solution, one that works off of the Pawn's _process function but CAN
be blocked or updated on request by KinematicDriverMachine.

The other issue is that the KinematicPawns seem to have an overtly
herky-jerky movement. You should watch them walk up stairs, it's
hilarious. Anyway, that's broken, but it's nothing that we can't patch
up with relative ease.