Update documentation to reflect new Pawns
The documentation needed to be updated to reflect the new varieities
of pawn. Their function, and their purpose. Most of the old
documentation still referred to the KinematicDriver.

I'm not gonna lie, this was like four hours of work stretched out over
three days because I was NOT feeling it. But documentation is
important, so I gotta stick with it.