Fix slope stuttering in the Kinematic Drive Machine
Previously, the Kinematic Driver Machine struggled when walking across
slopes. It did this weird stuttering thing and it DID make but it went
pretty slowly.

I managed to identify the issue was in the OnGround machine
superstate. Transitioning instantly to the fall state when a fall was
detected was breaking everything.

So, the solution was two fold. First, I copied over the floor
detection code from the old KinematicDriver (good thing we got git). I
then modified it slightly to reflect that it was a machine state. So,
we now do some slight falling in the OnGround state (as necessary).

Next, we added a timer to delay our transition to the fall state as
well as the supporting variables. This necessitated a new configurable
in the KinematicCore - Fall State Delay Time. This allows us to
manually fiddle with the delay between leaving the edge and entering
the hard fall state, per integrating scene. Neat!

Also, I had to add a note on the Max Slope Degrees configurable. It is
no longer happy when trying to climb a slope that is equal to it's
maximum slope. I now recommend a +- 1 degree buffer on that
configurable.