Use a state key string for animations, not a signal
Previously, we were updating our sprites via the direct intervention
of a signal that came from the Machines. This was causing frame
issues and other such lag becuase not only were we firing it many
times a second, but we were updating the Pawn's sprite each time.

I've removed the visual_update signal and renamed/revised the
readable_state variable into the state_key variable. In order to pick
the correct animation, you need to interpret the state_key
appropriately.

I thought this better aligned with our overall goal of having AI
machines that exist generally idependent of the bodies that implement
them.