Fix the Dynamic Nav Mesh Test scene
The Dynamic Nav Mesh Test scene was broken. You could order the Pawn
to move around but it wouldn't actually move - it would just kinda be
stuck. It'd be in walk mode, but the animation wouldn't go anywhere.

The problem was that the KinematicPawn was in "Floor" configuration
and didn't have an assigned NavMesh. Ergo, the Pawn was trying to walk
towards a goal it was simultaneously already at and could never reach.

Reconfiguring the Kinematic Pawn has improved the issue.