Fix the microposition Loop error in the KDM
The Microposition Loop error is a long-recognized problem that
originated with the KinematicDriver. This error occurs when a body
that has integrated the Kinematic Driver Machine (KDM) oversteps it's
target position, then reverses course and oversteps by the exact same
amount in the opposite direction. We actually have code in place to
counteract oversteps; this seems to happen as an interplay between our
configured speed, the physics engine, and the level geometry or
navigation mesh.

Anyway, this error hadn't been addressed yet for the Kinematic Driver
Machine. We were running into it pretty regularly, so I decided to
take the time to address it.

Our new solution is, in my opinion, an improvement over what we did
for the basal KinematicDriver. Everytime we think we detect the
positional error outlined above, we increase our goal tolerance by a
configured amount. Eventually, we cap out and send out a stuck signal.

I've found that this works really well for detecting actual instances
of being properly stuck while handling the microposition looping
probelm. Ergo, we have a new signal that indicates we're stuck. We may
eventually move to handling that in-state.

I decided to update the to-do list because, with this update, we've
managed to create a fully-fledged replacement for the KinematicDriver.
The last thing I want to do, before moving on to the Attack model, is
develop a new slope-step test (I'm rather done with the old one).