Add the Advanced Kinematics Test
While the Slope Step Test was good, it suffered from some pretty
serious handicaps. We couldn't skip or cancel tests, the tests were
constructed by hand (which was hard to modify), we couldn't modulate
the time at all, and it was difficult to observe the *Pawn* falling. I
became convinced that there MUST have been a better way.

The Advanced Kinematics Test is the sum of that effort. It features
three tests - the fall test, the slope test, and the step test. The
slope and the step test have configurations so we can test different
levels with ease. The fall test allows us to test how the Pawn falls.

I wanted to add control of the time/engine speed, so I separated off
the GUI we had in the Viewport Shader Test into it's own scene - the
Time GUI. That GUI element had existed by itself for a while so I
found it very easy to separate it off into it's own scene.

I also updated the README files appropriately.

Now, I've already observed some issues. First, the Pawn is not very
good at detecting when it is legitimately stuck. Secondly, it
sometimes gets stuck on the tower. Also, the tower conflicts with the
testing perogatives for slope and the step - I can't make the "agent
step" navigation mesh configurable too high or it breaks the tower
test. However, too low and it reduces the flexibility of the slope and
step tests. We'll need to do a redesign, for sure...