Revise the Advanced Kinematic fall test
The Advanced Kinematic fall test was problematic because it conflicted
with our step and slope tests. This was mostly due to how the dynamic
navigation mesh had to be set up. In order to get the Pawn to try (and
fail) to run into too-steep slopes and too-tall steps, we had to
increase the agent max step and max slope.

This worked for the slope and step but fell apart when applied to the
smaller and more delicate fall test and the odd tower I had designed.
The Pawn kept getting stuck because it was trying to take shortcuts
it literally COULD NOT take.

Then I realized - why was I making the Pawn walk up some steps when
I could literally teleport it into the sky and just let it fall? That
was a much better idea.

So, the Advanced Kinematic fall test now has an adjustable fall height
that changes how high up we drop the Pawn from. I thought the tower
was actually a good stress test of the Pawn's navigation abilities, as
well as of the Dynamic Navigation mesh's functionality. So, I
relocated it to the Dynamic Navigation Mesh test, alongside a control
to sink it or raise it up.

Also, updated the CameraRig. 0 was equivalent to -1 in that it allowed
for camera free movement (sans clamping). However, I wanted no
movement on a particular axis, but 0 didn't clamp. So I changed the
code to allow 0 to clamp.

We're still seeing that the Kinematic Driver Machine isn't detecting
all instances of an integrated body getting stuck. We should try and
sort that out...