Patch kinematic issues (mostly)
We've had a longstanding issue of the Pawn getting stuck in a stupid
loop. We added processing for when the loop was inconsequentially
small - everything else was supposed to be handled under an error
signal.

Except that the error signal wasn't working. There was a lot of
reasons for that. On one hand, the "stuck" errors we were trying to
detect were too big. The Pawn was covering too much ground for the
history size we had configured. The solution there was to increase
the size from 6 to 24. I'm worried about it being too sensitive, so
only time will tell...

Another reason was that the lengths we were measuring were too
precise. We had six figures following the decimal, of course we
weren't seeing repeats. So, I added a new configurable to specify the
level of rounding we do for history testing. This has shown much more
promising results.

And the final reason was... sometimes the level geometry is just bad.
It's just bad and it's never going to be perfect, so you just have to
work around it. Our error handling gives us flexibility, now we need
to design around the flexibility.