Clear out target position when getting a new path
The KinematicPawn had this funny behavior where it would move on an
old path for a little bit whenever it ot a new path. Then it would
follow the new path, frequently doubling back.

I realized this was because we never touched the target_position
variable, only the target_path. Ergo, the pawn was still heading on
part of it's old path.

So I updated it - setting the target_path now clears the
target_position variable. This forces the Kinematic Driver Machine to
follow the specified waypoint from moment one.