Add a new WIP machine, the Immobile Killer Machine
The next phase of development will be developing our attack model. As
part of that effort, I thought it would be prudent to develop a new
brain that is exclusively focused on attacking.

That's what the Immobile Killer Machine will be (eventually). A sorta
almost turret machine that only focused on scanning for targets and
attacking those targets in range.

Unfortunately, unlike the Kinematic Driver Machine, this is uncharted
territory. I really have no idea what I'm doing. Having a machine to
run the attacks is all well and good but I don't know what the attacks
look like or what they do.

So I'm comitting what I have now, pending later redevelopment when we
have a more stable concept of what the attack model is and what it
does.