Update the to do with EXCITING NEW TOOLS
While browsing the internet, I came across a couple of tools that can
really kick the project into high gear.

The first is PicoCAD, a computer-aided design (3D modeling) program
created using the incredibly basic Pico8 enginge. It makes designing
little polygonal 3D models STUPIDLY easy, using an interface similar
to Valve's Hammer Editor (YES!!!). It also generates UVs for the
models - meaning we can ACTUALLY do texturing.

The second is Qodot, which is a plugin to import Quake .map files.
That's neat in itself; however, when combined with the TenchBroom
Level editor (an open source and cross-platform map editor), it
essentially gives Godot the level-and-prefab editor that it really
needs. And, better yet, it's a Hammer style editor (Hammer being a
descendant of the original Quake level editor)!!!!

Also, I removed some of to do items that I felt we had already
accomplished in some way. Or completed in spirit, as it were.