Develop an Attack Model test
I tried developing the Immobile Killer Machine, but quickly discovered
that I didn't really know what I was doing. Like, at all. I was trying
to create states to address problems that didn't exist with systems I
hadn't designed yet.

I thought that developing the Attack Model Test would help me towards
formulating what the Attack Model was supposed to look like. I'm not
sure if it did, but now I have a nice Attack Model Test.

Right now, it's just two Pawns. One walks around, following a torus in
a circle, and the other sits at the middle. Eventually, the middle
will be replaced by whatever is doing the attacking. The size of the
torus is adjustable, so that we can (eventually) test multiple attacks
of varying ranges.

Honestly, without a developed attack model, it's hard for me to say if
this is even a good test or not. I'm gonna have to figure this out...